+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### Nomad Books 0.5.2 - 1.16.1
 - Changed firefly camp limit render to use custom particles
-
 - Fixed an issue where players would get kick from a server upon applying itinerant ink
+- Fixed an issue where some blocks would drop items upon retrieving the camp
 
 ### Nomad Books 0.5.1 - 1.16.1
 - Fixed an issue where players would get instantly kicked from a server upon joining